upload, but do not tag. just gonna keep a record before I delete the older and src_better

+++ src_old/HiBusServo.cpp
@@ -1,86 +1,40 @@
 #include "HiBusServo.h"
 
-/**
- * @brief Construct a new HiBusServo::HiBusServo object
- * 
- * @param debug Enable debug output
- * @param maxRetries Maximum number of retries for reading from servo
- */
+
 HiBusServo::HiBusServo(int maxRetries)
     : _maxRetries(maxRetries), _serialPort(nullptr)
 {
 }
 
-/**
- * @brief Initialize the serial port for communication with the servo
- * 
- * @param serial The serial port to use
- */
 void HiBusServo::begin(HardwareSerial &serial)
 {
     _serialPort = &serial;
     _serialPort->begin(115200);
 }
 
-/**
- * @brief Set the maximum number of retries for reading from the servo
- * 
- * @param retries The maximum number of retries
- */
 void HiBusServo::setMaxRetries(int retries)
 {
     _maxRetries = retries;
 }
 
-// Basic movement commands
-/**
- * @brief Move the servo to a specific position in a given time
- * 
- * @param servoId The ID of the servo
- * @param destination The destination angle (-120 to 120 degrees)
- * @param time The time to complete the movement in milliseconds
- */
-void HiBusServo::moveTo(int servoId, float destination, uint16_t time)
+void HiBusServo::moveTo(uint8_t servoId, float destination, uint16_t time)
 {
-    // Input is in angle relative to center as zero, from -120 to 120.
-    // Negative numbers clockwise (when looking at face of servo.)
     int16_t servo_position_units = static_cast<int16_t>((destination * (500.0 / 120.0)) + 500.0);
-    
-    // Set to servo mode first, then move
     _serial_servo_set_mode(servoId, 0, 0);
     _serial_servo_move(servoId, servo_position_units, time);
 }
 
-/**
- * @brief Spin the servo at a specific velocity
- * 
- * @param servoId The ID of the servo
- * @param velocity The velocity to spin at (-100 to 100)
- */
-void HiBusServo::spinAt(int servoId, float velocity)
+void HiBusServo::spinAt(uint8_t servoId, float velocity)
 {
-    // Input should be from -100 to 100, multiply by 10 to get -1000 to 1000 range
     _serial_servo_set_mode(servoId, 1, velocity * 10);
 }
 
-/**
- * @brief Stop the servo's movement
- * 
- * @param servoId The ID of the servo
- */
-void HiBusServo::stopMove(int servoId)
+void HiBusServo::stopMove(uint8_t servoId)
 {
     _serial_servo_stop_move(servoId);
 }
 
-// Position and status reading
-/**
- * @brief Read the current position of the servo
- * 
- * @param servoId The ID of the servo
- * @return int The current position of the servo, or -1 on failure
- */
-int HiBusServo::readPosition(int servoId)
+int16_t HiBusServo::readPosition(uint8_t servoId)
 {
     int position = -1;
     for (int i = 0; i < _maxRetries; ++i)
@@ -95,13 +49,7 @@ int HiBusServo::readPosition(int servoId)
     return position;
 }
 
-/**
- * @brief Read the input voltage of the servo
- * 
- * @param servoId The ID of the servo
- * @return int The input voltage in mV, or -2048 on failure
- */
-int HiBusServo::readVin(int servoId)
+int16_t HiBusServo::readVin(uint8_t servoId)
 {
     int vin = -2048;
     for (int i = 0; i < _maxRetries; ++i)
@@ -116,21 +64,15 @@ int HiBusServo::readVin(int servoId)
     return vin;
 }
 
-/**
- * @brief Read the temperature of the servo
- * 
- * @param servoId The ID of the servo
- * @return int The temperature in degrees Celsius, or -1 on failure
- */
-int HiBusServo::readTemperature(int servoId)
+int8_t HiBusServo::readTemperature(uint8_t servoId)
 {
-    int temperature = -1;
+    int8_t temperature = -1;
     for (int i = 0; i < _maxRetries; ++i)
     {
         uint8_t data[1];
         if (_read_servo_param(servoId, CMD_TEMP_READ, data, 1))
         {
-            temperature = data[0];
+            temperature = (int8_t)data[0];  
             break;
         }
         delayMicroseconds(500);
@@ -138,38 +80,17 @@ int HiBusServo::readTemperature(int servoId)
     return temperature;
 }
 
-// Servo configuration
-/**
- * @brief Set the ID of the servo
- * 
- * @param oldID The current ID of the servo
- * @param newID The new ID for the servo
- */
-void HiBusServo::setServoID(int oldID, int newID)
+void HiBusServo::setServoID(uint8_t oldID, uint8_t newID)
 {
     _serial_servo_set_id(oldID, newID);
 }
 
-/**
- * @brief Set the mode of the servo
- * 
- * @param servoId The ID of the servo
- * @param mode The mode to set (0 for position, 1 for wheel)
- * @param speed The speed for wheel mode
- */
-void HiBusServo::setServoMode(int servoId, uint8_t mode, int16_t speed)
+void HiBusServo::setServoMode(uint8_t servoId, uint8_t mode, int16_t speed)
 {
     _serial_servo_set_mode(servoId, mode, speed);
 }
 
-/**
- * @brief Set the angle limits for the servo
- * 
- * @param servoId The ID of the servo
- * @param minAngle The minimum angle limit
- * @param maxAngle The maximum angle limit
- */
-void HiBusServo::setAngleLimits(int servoId, int16_t minAngle, int16_t maxAngle)
+void HiBusServo::setAngleLimits(uint8_t servoId, int16_t minAngle, int16_t maxAngle)
 {
     uint8_t data[4];
     data[0] = _get_low_byte(minAngle);
@@ -179,14 +100,7 @@ void HiBusServo::setAngleLimits(int servoId, int16_t minAngle, int16_t maxAngle)
     _write_servo_param(servoId, CMD_ANGLE_LIMIT_WRITE, data, 4);
 }
 
-/**
- * @brief Set the input voltage limits for the servo
- * 
- * @param servoId The ID of the servo
- * @param minVin The minimum voltage limit in mV
- * @param maxVin The maximum voltage limit in mV
- */
-void HiBusServo::setVinLimits(int servoId, uint16_t minVin, uint16_t maxVin)
+void HiBusServo::setVinLimits(uint8_t servoId, uint16_t minVin, uint16_t maxVin)
 {
     uint8_t data[4];
     data[0] = _get_low_byte(minVin);
@@ -196,84 +110,40 @@ void HiBusServo::setVinLimits(int servoId, uint16_t minVin, uint16_t maxVin)
     _write_servo_param(servoId, CMD_VIN_LIMIT_WRITE, data, 4);
 }
 
-/**
- * @brief Set the maximum temperature limit for the servo
- * 
- * @param servoId The ID of the servo
- * @param maxTemp The maximum temperature limit in degrees Celsius
- */
-void HiBusServo::setTempMaxLimit(int servoId, uint8_t maxTemp)
+void HiBusServo::setTempMaxLimit(uint8_t servoId, uint8_t maxTemp)
 {
     _write_servo_param(servoId, CMD_TEMP_MAX_LIMIT_WRITE, &maxTemp, 1);
 }
 
-/**
- * @brief Set the angle offset for the servo
- * 
- * @param servoId The ID of the servo
- * @param offset The angle offset
- */
-void HiBusServo::setAngleOffset(int servoId, int8_t offset)
+void HiBusServo::setAngleOffset(uint8_t servoId, int8_t offset)
 {
     uint8_t data = static_cast<uint8_t>(offset);
     _write_servo_param(servoId, CMD_ANGLE_OFFSET_WRITE, &data, 1);
 }
 
-// Servo control
-/**
- * @brief Enable the servo's torque
- * 
- * @param servoId The ID of the servo
- */
-void HiBusServo::loadServo(int servoId)
+void HiBusServo::loadServo(uint8_t servoId)
 {
     _serial_servo_load(servoId);
 }
 
-/**
- * @brief Disable the servo's torque
- * 
- * @param servoId The ID of the servo
- */
-void HiBusServo::unloadServo(int servoId)
+void HiBusServo::unloadServo(uint8_t servoId)
 {
     _serial_servo_unload(servoId);
 }
 
-/**
- * @brief Set the servo's LED state
- * 
- * @param servoId The ID of the servo
- * @param ledOn True to turn the LED on, false to turn it off
- */
-void HiBusServo::setLEDControl(int servoId, bool ledOn)
+void HiBusServo::setLEDControl(uint8_t servoId, bool ledOn)
 {
     uint8_t data = ledOn ? 1 : 0;
     _write_servo_param(servoId, CMD_LED_CTRL_WRITE, &data, 1);
 }
 
-/**
- * @brief Set the servo's LED error indicator state
- * 
- * @param servoId The ID of the servo
- * @param errorLedOn True to turn the error LED on, false to turn it off
- */
-void HiBusServo::setLEDErrorControl(int servoId, bool errorLedOn)
+void HiBusServo::setLEDErrorControl(uint8_t servoId, bool errorLedOn)
 {
     uint8_t data = errorLedOn ? 1 : 0;
     _write_servo_param(servoId, CMD_LED_ERROR_WRITE, &data, 1);
 }
 
-// Getters for servo parameters
-/**
- * @brief Get the servo's mode
- * 
- * @param servoId The ID of the servo
- * @param mode Variable to store the mode
- * @param speed Variable to store the speed
- * @return true if successful, false otherwise
- */
-bool HiBusServo::getServoMode(int servoId, uint8_t &mode, int16_t &speed)
+bool HiBusServo::getServoMode(uint8_t servoId, uint8_t &mode, int16_t &speed)
 {
     for (int i = 0; i < _maxRetries; ++i)
     {
@@ -289,15 +159,7 @@ bool HiBusServo::getServoMode(int servoId, uint8_t &mode, int16_t &speed)
     return false;
 }
 
-/**
- * @brief Get the servo's angle limits
- * 
- * @param servoId The ID of the servo
- * @param minAngle Variable to store the minimum angle
- * @param maxAngle Variable to store the maximum angle
- * @return true if successful, false otherwise
- */
-bool HiBusServo::getAngleLimits(int servoId, int16_t &minAngle, int16_t &maxAngle)
+bool HiBusServo::getAngleLimits(uint8_t servoId, int16_t &minAngle, int16_t &maxAngle)
 {
     for (int i = 0; i < _maxRetries; ++i)
     {
@@ -313,15 +175,7 @@ bool HiBusServo::getAngleLimits(int servoId, int16_t &minAngle, int16_t &maxAngl
     return false;
 }
 
-/**
- * @brief Get the servo's input voltage limits
- * 
- * @param servoId The ID of the servo
- * @param minVin Variable to store the minimum voltage
- * @param maxVin Variable to store the maximum voltage
- * @return true if successful, false otherwise
- */
-bool HiBusServo::getVinLimits(int servoId, uint16_t &minVin, uint16_t &maxVin)
+bool HiBusServo::getVinLimits(uint8_t servoId, uint16_t &minVin, uint16_t &maxVin)
 {
     for (int i = 0; i < _maxRetries; ++i)
     {
@@ -337,14 +191,7 @@ bool HiBusServo::getVinLimits(int servoId, uint16_t &minVin, uint16_t &maxVin)
     return false;
 }
 
-/**
- * @brief Get the servo's maximum temperature limit
- * 
- * @param servoId The ID of the servo
- * @param maxTemp Variable to store the maximum temperature
- * @return true if successful, false otherwise
- */
-bool HiBusServo::getTempMaxLimit(int servoId, uint8_t &maxTemp)
+bool HiBusServo::getTempMaxLimit(uint8_t servoId, uint8_t &maxTemp)
 {
     for (int i = 0; i < _maxRetries; ++i)
     {
@@ -359,14 +206,7 @@ bool HiBusServo::getTempMaxLimit(int servoId, uint8_t &maxTemp)
     return false;
 }
 
-/**
- * @brief Get the servo's angle offset
- * 
- * @param servoId The ID of the servo
- * @param offset Variable to store the angle offset
- * @return true if successful, false otherwise
- */
-bool HiBusServo::getAngleOffset(int servoId, int8_t &offset)
+bool HiBusServo::getAngleOffset(uint8_t servoId, int8_t &offset)
 {
     for (int i = 0; i < _maxRetries; ++i)
     {
@@ -381,14 +221,7 @@ bool HiBusServo::getAngleOffset(int servoId, int8_t &offset)
     return false;
 }
 
-/**
- * @brief Get the servo's LED state
- * 
- * @param servoId The ID of the servo
- * @param ledOn Variable to store the LED state
- * @return true if successful, false otherwise
- */
-bool HiBusServo::getLEDControl(int servoId, bool &ledOn)
+bool HiBusServo::getLEDControl(uint8_t servoId, bool &ledOn)
 {
     for (int i = 0; i < _maxRetries; ++i)
     {
@@ -403,14 +236,7 @@ bool HiBusServo::getLEDControl(int servoId, bool &ledOn)
     return false;
 }
 
-/**
- * @brief Get the servo's LED error indicator state
- * 
- * @param servoId The ID of the servo
- * @param errorLedOn Variable to store the LED error state
- * @return true if successful, false otherwise
- */
-bool HiBusServo::getLEDErrorControl(int servoId, bool &errorLedOn)
+bool HiBusServo::getLEDErrorControl(uint8_t servoId, bool &errorLedOn)
 {
     for (int i = 0; i < _maxRetries; ++i)
     {
@@ -425,13 +251,6 @@ bool HiBusServo::getLEDErrorControl(int servoId, bool &errorLedOn)
     return false;
 }
 
-// Private helper methods
-/**
- * @brief Calculate the checksum for a command packet
- * 
- * @param buf The buffer containing the command packet
- * @return uint8_t The calculated checksum
- */
 uint8_t HiBusServo::_calculate_checksum(uint8_t buf[])
 {
     uint16_t temp = 0;
@@ -443,19 +262,11 @@ uint8_t HiBusServo::_calculate_checksum(uint8_t buf[])
     return (uint8_t)temp;
 }
 
-/**
- * @brief Send a move command to the servo
- * 
- * @param id The ID of the servo
- * @param position The target position (0-1000)
- * @param time The time for the movement in milliseconds
- */
 void HiBusServo::_serial_servo_move(uint8_t id, int16_t position, uint16_t time)
 {
     uint8_t buf[10];
     if (position < 0) position = 0;
     if (position > 1000) position = 1000;
-    
     buf[0] = buf[1] = FRAME_HEADER;
     buf[2] = id;
     buf[3] = 7;
@@ -468,11 +279,6 @@ void HiBusServo::_serial_servo_move(uint8_t id, int16_t position, uint16_t time)
     _serialPort->write(buf, 10);
 }
 
-/**
- * @brief Send a stop move command to the servo
- * 
- * @param id The ID of the servo
- */
 void HiBusServo::_serial_servo_stop_move(uint8_t id)
 {
     uint8_t buf[6];
@@ -484,12 +290,6 @@ void HiBusServo::_serial_servo_stop_move(uint8_t id)
     _serialPort->write(buf, 6);
 }
 
-/**
- * @brief Send a set ID command to the servo
- * 
- * @param oldID The current ID of the servo
- * @param newID The new ID for the servo
- */
 void HiBusServo::_serial_servo_set_id(uint8_t oldID, uint8_t newID)
 {
     uint8_t buf[7];
@@ -502,13 +302,6 @@ void HiBusServo::_serial_servo_set_id(uint8_t oldID, uint8_t newID)
     _serialPort->write(buf, 7);
 }
 
-/**
- * @brief Send a set mode command to the servo
- * 
- * @param id The ID of the servo
- * @param mode The mode to set
- * @param speed The speed for wheel mode
- */
 void HiBusServo::_serial_servo_set_mode(uint8_t id, uint8_t mode, int16_t speed)
 {
     uint8_t buf[10];
@@ -524,11 +317,6 @@ void HiBusServo::_serial_servo_set_mode(uint8_t id, uint8_t mode, int16_t speed)
     _serialPort->write(buf, 10);
 }
 
-/**
- * @brief Send a load (enable torque) command to the servo
- * 
- * @param id The ID of the servo
- */
 void HiBusServo::_serial_servo_load(uint8_t id)
 {
     uint8_t buf[7];
@@ -541,11 +329,6 @@ void HiBusServo::_serial_servo_load(uint8_t id)
     _serialPort->write(buf, 7);
 }
 
-/**
- * @brief Send an unload (disable torque) command to the servo
- * 
- * @param id The ID of the servo
- */
 void HiBusServo::_serial_servo_unload(uint8_t id)
 {
     uint8_t buf[7];
@@ -558,13 +341,7 @@ void HiBusServo::_serial_servo_unload(uint8_t id)
     _serialPort->write(buf, 7);
 }
 
-/**
- * @brief Handle receiving data from the servo
- * 
- * @param ret Buffer to store the received data
- * @return int 1 on success, -1 on checksum error, 0 if no data
- */
-int HiBusServo::_serial_servo_receive_handle(uint8_t *ret)
+int16_t HiBusServo::_serial_servo_receive_handle(uint8_t *ret)
 {
     bool frameStarted = false;
     bool receiveFinished = false;
@@ -625,13 +402,7 @@ int HiBusServo::_serial_servo_receive_handle(uint8_t *ret)
     return 0;
 }
 
-/**
- * @brief Read the position from the servo
- * 
- * @param id The ID of the servo
- * @return int The position, or -1 on failure
- */
-int HiBusServo::_serial_servo_read_position(uint8_t id)
+int16_t HiBusServo::_serial_servo_read_position(uint8_t id)
 {
     int count = 10000;
     int ret;
@@ -669,13 +440,7 @@ int HiBusServo::_serial_servo_read_position(uint8_t id)
     return ret;
 }
 
-/**
- * @brief Read the input voltage from the servo
- * 
- * @param id The ID of the servo
- * @return int The input voltage in mV, or -2048 on failure
- */
-int HiBusServo::_serial_servo_read_vin(uint8_t id)
+int16_t HiBusServo::_serial_servo_read_vin(uint8_t id)
 {
     int count = 10000;
     int ret;
@@ -713,15 +478,6 @@ int HiBusServo::_serial_servo_read_vin(uint8_t id)
     return ret;
 }
 
-// Generic read/write methods for new functionality
-/**
- * @brief Generic method to write a parameter to the servo
- * 
- * @param id The ID of the servo
- * @param cmd The command to send
- * @param data The data to send
- * @param dataLen The length of the data
- */
 void HiBusServo::_write_servo_param(uint8_t id, uint8_t cmd, uint8_t *data, uint8_t dataLen)
 {
     uint8_t buf[32];
@@ -739,15 +495,6 @@ void HiBusServo::_write_servo_param(uint8_t id, uint8_t cmd, uint8_t *data, uint
     _serialPort->write(buf, 6 + dataLen);
 }
 
-/**
- * @brief Generic method to read a parameter from the servo
- * 
- * @param id The ID of the servo
- * @param cmd The command to send
- * @param data Buffer to store the received data
- * @param expectedLen The expected length of the data
- * @return true if successful, false otherwise
- */
 bool HiBusServo::_read_servo_param(uint8_t id, uint8_t cmd, uint8_t *data, uint8_t expectedLen)
 {
     uint8_t buf[32];
